[AN] refactor: 검색 관련 Fragment 이름 변경 및 카테고리 화면 구조 개선 #755
*   [AN] refactor: SearchComposeFragment 및 CategoryComposeFragment 클래스명에서 Compose 제거
*   [AN] refactor: CategorySearchScreen 내부 컴포넌트(SearchedHearitItem, CustomLinearProgressBar) 파일 분리
*   [AN] refactor: CategorySearchScreen을 Stateless 구조로 변경 및 패키지 이동
