+++ android/app/src/main/java/com/onair/hearit/presentation/home/HomeFragment.kt
@@ -47,7 +47,7 @@ import com.onair.hearit.presentation.home.adapter.RecommendHearitAdapter
 import com.onair.hearit.presentation.home.adapter.RecommendationCategoryAdapter
 import com.onair.hearit.presentation.main.MainActivity
 import com.onair.hearit.presentation.main.MainViewModel
-import com.onair.hearit.presentation.search.category.CategoryComposeFragment
+import com.onair.hearit.presentation.search.category.CategoryFragment
 import kotlinx.coroutines.launch
 
 class HomeFragment :
@@ -339,7 +339,7 @@ class HomeFragment :
             .beginTransaction()
             .replace(
                 R.id.fragment_container_view,
-                CategoryComposeFragment().apply {
+                CategoryFragment().apply {
                     arguments =
                         bundleOf(
                             CATEGORY_ID_KEY to id,
