+++ android/app/src/main/java/com/onair/hearit/presentation/search/main/SearchFragment.kt
@@ -17,11 +17,11 @@ import com.onair.hearit.presentation.IntentKeys
 import com.onair.hearit.presentation.search.CategoryClickListener
 import com.onair.hearit.presentation.search.SearchViewModel
 import com.onair.hearit.presentation.search.SearchViewModelFactory
-import com.onair.hearit.presentation.search.category.CategoryComposeFragment
+import com.onair.hearit.presentation.search.category.CategoryFragment
 import com.onair.hearit.presentation.search.main.screen.SearchMainScreen
 import com.onair.hearit.presentation.search.recent.SearchRecentFragment
 
-class SearchComposeFragment :
+class SearchFragment :
     Fragment(),
     CategoryClickListener {
     private val viewModel: SearchViewModel by viewModels {
@@ -69,7 +69,7 @@ class SearchComposeFragment :
         colorCode: String,
     ) {
         val fragment =
-            CategoryComposeFragment().apply {
+            CategoryFragment().apply {
                 arguments =
                     bundleOf(
                         IntentKeys.CATEGORY_ID_KEY to id,
